feat(amount-shares): re-calculate amount from shares with BigInt
