+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "lambda-handlers-image",
-  "version": "0.8.6",
+  "version": "0.8.7",
   "private": true,
   "type": "module",
   "scripts": {
